Refactored routes to make more sense with the header and home page, added redux-promise back in to load events, fixed appbar to load on every page properly
