+++ client/src/App.js
@@ -1,16 +1,15 @@
 import React, { Component } from "react";
-import { connect } from 'react-redux';
+import { connect } from "react-redux";
 import Grid from "@material-ui/core/Grid";
 import { BrowserRouter as Router, Route, Switch } from "react-router-dom";
-
-import Login from "./components/Login";
+import Header from "./components/Header";
 import HomePage from "./components/HomePage";
+import EventPage from "./components/EventPage";
 import UserGroup from "./components/UserGroup";
 import NoMatch from "./containers/NoMatch";
-import * as actions from './actions/index';
+import * as actions from "./actions/index";
 
 class App extends Component {
-
   componentDidMount() {
     this.props.fetchGoogleUser();
   }
@@ -31,12 +30,18 @@ class App extends Component {
         <Grid container justify="center">
           <Grid item xs={12} md={10}>
             <Router>
-              <Switch>
-                <Route exact path="/" component={Login} />
-                <Route exact path="/events" component={HomePage} />
-                <Route path="/events/:event/:usergroupid" component={UserGroup} />
-                <Route component={NoMatch} />
-              </Switch>
+              <div>
+                <Header />
+                <Switch>
+                  <Route exact path="/" component={HomePage} />
+                  <Route exact path="/events" component={EventPage} />
+                  <Route
+                    path="/events/:event/:usergroupid"
+                    component={UserGroup}
+                  />
+                  <Route component={NoMatch} />
+                </Switch>
+              </div>
             </Router>
           </Grid>
         </Grid>
@@ -45,4 +50,7 @@ class App extends Component {
   }
 }
 
-export default connect(null, actions)(App);
+export default connect(
+  null,
+  actions
+)(App);
